Conflicts:
	Algorithm_Experiments/morphologic_process.m

+++ Algorithm_Experiments/rank_filter.m
@@ -2,5 +2,5 @@ function R = rank_filter(V, n)
     template = ones(1,n) / n;
     R = conv(V, template);
     [r, c] = size(R);
-    R = R(1, n/2: c - n/2);
+    R = R(1, int32(n/2): c - int32(n/2));
 end
